refactor: CartContainer and split into components
