+++ src/refactoring/components/CartPage.tsx
@@ -1,6 +1,6 @@
 import { Coupon, Product } from "../../types.ts";
 import { CartProvider } from "../contexts/CartContext.tsx";
-import { CartContainer } from "./CartList.tsx";
+import { CartContainer } from "./CartContainer.tsx";
 import { ProductList } from "./ProductList.tsx";
 
 interface Props {
